+++ modules/Settings/Requests/Currencies/UpdateCurrencyRequest.php
@@ -10,6 +10,7 @@ use Illuminate\Foundation\Http\FormRequest;
 
 final class UpdateCurrencyRequest extends FormRequest
 {
+
     /**
      * Get the validation rules that apply to the request.
      *
@@ -18,7 +19,7 @@ final class UpdateCurrencyRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'exchange' => ['required', 'float'],
+            'exchange' => ['required', 'numeric'],
             'symbol' => ['required', 'string', 'max:100'],
             'is_primary' => ['required', 'boolean'],
             'currency_code' => ['nullable', 'string', 'max:100'],
